fix build

+++ app/src/main/java/com/example/b07group7project/shopper_view_previous_orders/OrderedProduct.java
@@ -33,7 +33,7 @@ public class OrderedProduct {
         return product;
     }
 
-    public int getQuantity() {
+    public long getQuantity() {
         return quantity;
     }
